refactor(ka-routes): Stage 1.2 — extract shared assertion helpers into shared-assertion-helpers.ts
Moves 30+ helpers (import-artifact resolution, owner-guard, semantic-enrichment
RDF builders, promote-async job views, quad sort) out of assertion.ts into a new
shared module so the new /api/knowledge-assets route ports can reuse them and
assertion.ts can be deleted in Stage 4. assertion.ts re-imports them; behavior
unchanged. Typecheck-clean.
